refactor: move major parts of the code
Move the actual code from telescope/_extension folder into the actual
plugin folder.
