+++ lua/telescope-orgmode/init.lua
@@ -1,0 +1,4 @@
+return {
+  refile_heading = require('telescope-orgmode.refile_heading'),
+  search_headings = require('telescope-orgmode.search_headings'),
+}
